+++ CHANGELOG.md
@@ -9,6 +9,7 @@ The most recent changes are on top, in each type of changes category.
 
 ### Added
 
+- Query param handling in search endpoint [#512](https://github.com/cyfronet-fid/sat4envi/issues/512)
 - Add keys handling for search [#556](https://github.com/cyfronet-fid/sat4envi/issues/556)
 
 ### Changed